Removed files no longer required. Changed function name for displaying how much a file has been loaded for preview. Changed border constant name.
