--- src/main/webapp/pages/analysis/details/unavailable.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      data-layout-decorate="~{analysis/details/details}">
-
-<body>
-  <th:block layout:fragment="noPreview">
-    <h2 th:text="#{analysis.details.preview.unavailable}">_Preview not available_</h2>
-  </th:block>
-
-</body>
-
-</html>
