+++ src/main/webapp/resources/js/components/OutputFiles/JsonOutputWrapper.jsx
@@ -6,25 +6,34 @@ import React from "react";
 import PropTypes from "prop-types";
 import { SPACE_XS } from "../../styles/spacing";
 import styled from "styled-components";
-import { border1 } from "../../styles/borders";
+import { DIV_BORDER } from "../../styles/borders";
 
 /**
  * Stateless UI component for displaying a div for json output
  *
  * @param {number} height - Height of div. Default is 300px
- * @param {string} description - Optional description
+ * @param {bool} overflowRequired - Optional overflow for div
  * @param {object} children - content to display in div
  * @param {object} props - any other attributes to add to JsonOutputWrapper
  *
  * @returns {Element} - Returns a scrollable div component
  */
 
-export function JsonOutputWrapper({ height = 300, children, ...props }) {
+export function JsonOutputWrapper({
+  height = 300,
+  overflowRequired = false,
+  children,
+  ...props
+}) {
+  const overflow = overflowRequired ? "auto" : "none";
+
   const Wrapper = styled.div`
     height: ${height}px;
     margin-bottom: ${SPACE_XS};
-    border: ${border1};
+    border: ${DIV_BORDER};
     padding: 0 0 2px ${SPACE_XS};
+    white-space: pre;
+    overflow: ${overflow};
   `;
 
   return (
@@ -37,8 +46,14 @@ export function JsonOutputWrapper({ height = 300, children, ...props }) {
 JsonOutputWrapper.propTypes = {
   /*Optional height*/
   height: PropTypes.number,
+  /*Optional overflowRequired*/
+  overflowRequired: PropTypes.bool,
   /*the content to display in the div*/
-  children: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
+  children: PropTypes.oneOfType([
+    PropTypes.object,
+    PropTypes.array,
+    PropTypes.string
+  ]),
   /*any extra attributes to add to JsonOutputWrapper*/
   props: PropTypes.object
 };
