+++ src/main/webapp/resources/js/components/OutputFiles/index.js
@@ -1,3 +1,2 @@
-export { JsonObjectOutputWrapper } from "./JsonObjectOutputWrapper";
 export { JsonOutputWrapper } from "./JsonOutputWrapper";
 export { OutputFileHeader } from "./OutputFileHeader";
