+++ src/main/webapp/resources/js/pages/samples/sample-files.js
@@ -1,7 +1,7 @@
 const angular = require("angular");
 import "angular-ui-bootstrap";
 require("ng-file-upload");
-import { convertFileSize } from "../../utilities/file.utilities";
+import { convertFileSize } from "../../utilities/file-utilities";
 require("../../../sass/pages/sample-files.scss");
 
 /**
