+++ src/main/webapp/resources/js/styles/borders.js
@@ -7,4 +7,4 @@ import { grey4 } from "./colors";
  * LEAD ON INTERFACE DESIGN.
  */
 
-export const border1 = `solid 1px ${grey4}`;
+export const DIV_BORDER = `solid 1px ${grey4}`;
