+++ src/main/webapp/resources/js/utilities/file-utilities.js
@@ -32,3 +32,28 @@ export function convertFileSize(bytes) {
   } while (Math.abs(bytes) >= thresh && u < units.length - 1);
   return bytes.toFixed(1) + " " + units[u];
 }
+
+/**
+ *
+ * Returns how much of a file has been loaded into preview div
+ * @param {number} byte Number of bytes of file currently fetched from server
+ * @param {number} fileSizeBytes File size in bytes
+ * @returns {string} Returns file size loaded (x of y bytes as well as the percentage of the file)
+ */
+export function fileSizeLoaded(byte, fileSizeBytes) {
+  return `${convertFileSize(byte)} / ${convertFileSize(fileSizeBytes)} (${(
+    (byte / fileSizeBytes) *
+    100
+  ).toFixed(1)}%)`;
+}
+
+/**
+ *
+ * @param {number} filePosition Starting byte in file
+ * @param {number} fileSizeBytes Total file size
+ * @param {number} chunkSize Number of bytes to retrieve from filePosition
+ * @returns {number} Size of chunk to retrieve from file in bytes
+ */
+export function getNewChunkSize(filePosition, fileSizeBytes, chunkSize) {
+  return Math.min(fileSizeBytes - filePosition, chunkSize);
+}
